requirements.txt
# Core Dependencies
tensorflow==2.15.0
keras==2.15.0
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.1.0
scipy==1.11.4

# Web Framework
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

# Image Processing
scikit-image==0.22.0
imageio==2.31.5

# Data Processing
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0

# Machine Learning Utilities
scikit-learn==1.3.2
imutils==0.5.4

# API and Utilities
requests==2.31.0
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==23.11.0
flake8==6.1.0
pylint==3.0.3

# Progress bars
tqdm==4.66.1

# XML Processing (for annotations)
lxml==4.9.3
